Create domain folder, move modules from init
Refs: https://github.com/HowProgrammingWorks/NodejsStarterKit/issues/84
