+++ init/logResources.js
@@ -1,19 +1,9 @@
 'use strict';
 
-const v8 = require('v8');
-
-const getStatistics = () => {
-  const { heapTotal, heapUsed, external } = process.memoryUsage();
-  const hs = v8.getHeapStatistics();
-  const contexts = hs.number_of_native_contexts;
-  const detached = hs.number_of_detached_contexts;
-  return { heapTotal, heapUsed, external, contexts, detached };
-};
-
 const startMonitoring = application => {
   const config = application.config.sections.resmon;
   setInterval(() => {
-    const stats = getStatistics();
+    const stats = application.resmon.getStatistics();
     const { heapTotal, heapUsed, external, contexts, detached } = stats;
     const total = application.utils.bytesToSize(heapTotal);
     const used = application.utils.bytesToSize(heapUsed);
@@ -23,4 +13,4 @@ const startMonitoring = application => {
   }, config.interval);
 };
 
-module.exports = { getStatistics, startMonitoring };
+module.exports = startMonitoring;
